Use length-based packet separation (#3)
Before this patch, sending multiline text or binary files would have been
very difficult and inefficient as any newline they contained would get
interpreted as a different packet. Now, any communications are prefixed
by a number which denotes the size of the message which is going to be
sent subsequently.